app/src/androidTest/java/com/gunnarro/android/terex/service/ClientServiceTest.java
package com.gunnarro.android.terex.service;

import static org.junit.Assert.assertEquals;
import static org.junit.Assert.assertNotNull;
import static org.junit.Assert.assertNull;
import static org.junit.Assert.assertTrue;

import android.content.Context;

import androidx.room.Room;
import androidx.test.core.app.ApplicationProvider;

import com.gunnarro.android.terex.DbHelper;
import com.gunnarro.android.terex.config.AppDatabase;
import com.gunnarro.android.terex.domain.dto.ClientDto;
import com.gunnarro.android.terex.domain.dto.OrganizationDto;
import com.gunnarro.android.terex.domain.entity.Client;
import com.gunnarro.android.terex.repository.ClientRepository;
import com.gunnarro.android.terex.repository.ProjectRepository;

import org.junit.Before;
import org.junit.Ignore;
import org.junit.Test;

import java.util.List;

public class ClientServiceTest {

    private ClientService clientService;

    @Before
    public void setup() {
        Context appContext = ApplicationProvider.getApplicationContext();
        AppDatabase appDatabase = Room.inMemoryDatabaseBuilder(appContext, AppDatabase.class).build();
        AppDatabase.init(appContext);
        clientService = new ClientService(new ClientRepository(), new ProjectRepository(), new PersonService());
        // load test data
        List<String> sqlQueryList = DbHelper.readMigrationSqlQueryFile(appContext, "database/test_data.sql");
        sqlQueryList.forEach(query -> {
            System.out.println("DB test data sql query: : " + query);
            appDatabase.getOpenHelper().getWritableDatabase().execSQL(query);
        });
        assertTrue(appDatabase.getOpenHelper().getWritableDatabase().isDatabaseIntegrityOk());
    }

    /* fixme
    @Test
    public void getClient() {
        ClientDto clientDto = clientService.getClient(1L);
        assertNotNull(clientDto);
        assertEquals("GUNNARRO AS", clientDto.getName());
    }
*/
    @Test
    public void newAndUpdateClient() {
        ClientDto newClientDto = new ClientDto();
        OrganizationDto organizationDto = new OrganizationDto();
        organizationDto.setName("gunnarro as");
        organizationDto.setOrganizationNumber("822707922");
        newClientDto.setOrganizationDto(organizationDto);
        newClientDto.setName(organizationDto.getName());
        newClientDto.setStatus(Client.ClientStatusEnum.ACTIVE.name());
        Long id = clientService.saveClient(newClientDto);
        assertNotNull(id);
        // check new
        ClientDto clientDto = clientService.getClient(id);
        assertEquals(id, clientDto.getId());
        assertEquals("gunnarro as", clientDto.getName());
        assertEquals("ACTIVE", clientDto.getStatus());
        assertNull(clientDto.getOrganizationDto()); // fixme should be returned by get client
       // assertNotNull(clientDto.getOrganizationDto().getId());
       // assertEquals("822707922", clientDto.getOrganizationDto().getOrganizationNumber());
        // update client status
        clientDto.setStatus(Client.ClientStatusEnum.DEACTIVATED.name());
        id = clientService.saveClient(clientDto);
        // check updated
        ClientDto updatedClientDto = clientService.getClient(id);
        assertEquals("gunnarro as", updatedClientDto.getName());
        assertEquals("DEACTIVATED", updatedClientDto.getStatus());
    }
}
